requirements.txt
# Web Framework 
fastapi==0.103.1 
uvicorn==0.23.2 
python-multipart==0.0.6 
jinja2==3.1.2 
 
# Data Processing 
pandas==2.1.0 
numpy==1.25.2 
scikit-learn==1.3.0 
openpyxl==3.1.2 
 
matplotlib==3.7.2 
seaborn==0.12.2 
fpdf2==2.7.4 
pillow==10.0.0 
 
# Utilities 
python-dotenv==1.0.0 
